attritbute and entry timeout values under volume now support float/double values

+++ libglusterfs/src/glusterfs.h
@@ -225,8 +225,8 @@ struct _cmd_args {
 
 	/* fuse options */
 	int              fuse_direct_io_mode_flag;
-	unsigned int     fuse_entry_timeout;
-	unsigned int     fuse_attribute_timeout;
+	double           fuse_entry_timeout;
+	double           fuse_attribute_timeout;
 	char            *volume_name;
 	int              non_local;       /* Used only by darwin os, 
 					     used for '-o local' option */
